+++ node_helper.js
@@ -1,117 +1,172 @@
-/* Magic Mirror
- * Module: MMM-WienerLinien
+/**
+ * @file node_helper.js
  *
- * By fewieden https://github.com/fewieden/MMM-WienerLinien
- * MIT Licensed.
+ * @author fewieden
+ * @license MIT
+ *
+ * @see  https://github.com/fewieden/MMM-WienerLinien
  */
 
-/* eslint-env node */
+/**
+ * @external node-fetch
+ * @see https://www.npmjs.com/package/node-fetch
+ */
+const fetch = require('node-fetch');
 
-const request = require('request');
+/**
+ * @external node_helper
+ * @see https://github.com/MichMich/MagicMirror/blob/master/js/node_helper.js
+ */
 const NodeHelper = require('node_helper');
 
+/**
+ * @module node_helper
+ * @description Backend for the module to query data from the API provider.
+ *
+ * @requires external:node-fetch
+ * @requires external:node_helper
+ */
 module.exports = NodeHelper.create({
+    baseUrl: 'https://www.wienerlinien.at/ogd_realtime',
 
-    baseUrl: 'https://www.wienerlinien.at/ogd_realtime/',
-
-    start() {
-        console.log(`Starting module helper: ${this.name}`);
-    },
-
+    /**
+     * @function socketNotificationReceived
+     * @description Receives socket notifications from the module.
+     * @async
+     * @override
+     *
+     * @param {string} notification - Notification name
+     * @param {*} payload - Detailed payload of the notification.
+     *
+     * @returns {void}
+     */
     socketNotificationReceived(notification, payload) {
         if (notification === 'CONFIG') {
             this.config = payload;
             this.getData();
-            setInterval(() => {
-                this.getData();
+            setInterval(async () => {
+                await this.getData();
             }, this.config.updateInterval);
         }
     },
 
-    getData() {
-        const options = {
-            url: `${this.baseUrl}monitor?sender=${this.config.api_key}&rbl=${this.config.stations.join('&rbl=')}`
-        };
+    /**
+     * @function getData
+     * @description Wrapper for all required data set requests.
+     * @async
+     *
+     * @returns {void}
+     */
+    async getData() {
+        await this.getMonitoringData();
 
-        request(options, (error, response, body) => {
-            if (response.statusCode === 200) {
-                const parsedBody = JSON.parse(body);
-                if (parsedBody.message.value === 'OK') {
-                    this.handleData(parsedBody.data.monitors, parsedBody.message.serverTime);
-                } else {
-                    console.log('Error no WienerLinen data');
-                }
-            } else {
-                console.log(`Error getting WienerLinen data ${response.statusCode}`);
-            }
-        });
-
-        // Get elevator info
         if (this.config.elevatorStations.length > 0) {
-            options.url = `${this.baseUrl}trafficInfoList?sender=${this.config.api_key}&name=aufzugsinfo&relatedStop=${this.config.elevatorStations.join('&relatedStop=')}`;
+            const url = `${this.baseUrl}/trafficInfoList?name=aufzugsinfo&relatedStop=${this.config.elevatorStations.join('&relatedStop=')}`;
 
-            this.getAdditionalData(options, 'Elevator');
+            await this.getAdditionalData(url, 'Elevator');
         }
 
-        // Get incident info
         if (this.config.incidentLines.length > 0) {
-            let type = '&name=stoerunglang';
-            if (this.config.incidentShort) {
-                type += '&name=stoerungkurz';
-            }
+            const incidentType = this.config.incidentShort ? 'stoerungkurz' : 'stoerunglang';
+            const url = `${this.baseUrl}/trafficInfoList?name=${incidentType}&relatedLine=${this.config.incidentLines.join('&relatedLine=')}`;
 
-            options.url = `${this.baseUrl}trafficInfoList?sender=${this.config.api_key}${type}&relatedLine=${this.config.incidentLines.join('&relatedLine=')}`;
-
-            this.getAdditionalData(options, 'Incident');
+            await this.getAdditionalData(url, 'Incident');
         }
     },
 
-    getAdditionalData(options, type) {
-        request(options, (error, response, body) => {
-            if (response.statusCode === 200) {
-                const parsedBody = JSON.parse(body);
-                if (Object.prototype.hasOwnProperty.call(parsedBody.data, 'trafficInfos')) {
-                    this[`handle${type}Data`](parsedBody.data.trafficInfos);
-                } else {
-                    console.log(`Info: no WienerLinen ${type} data available.`);
-                }
+    /**
+     * @function getMonitoringData
+     * @description Requests departures of stations from monitoring service.
+     * @async
+     *
+     * @returns {void}
+     */
+    async getMonitoringData() {
+        try {
+            const url = `${this.baseUrl}/monitor?rbl=${this.config.stations.join('&rbl=')}`;
+
+            const response = await fetch(`${this.baseUrl}/monitor?rbl=${this.config.stations.join('&rbl=')}`);
+            const parsedBody = await response.json();
+
+            if (parsedBody.message.value === 'OK') {
+                this.handleData(parsedBody.data.monitors, parsedBody.message.serverTime);
             } else {
-                console.log(`Error getting WienerLinen ${type} data ${response.statusCode}`);
+                throw new Error('No WienerLinien data');
             }
-        });
+        } catch (e) {
+            console.log('Error getting WienerLinien station data:', e);
+        }
     },
 
-    handleElevatorData(data) {
-        const elevators = [];
-        for (let i = 0; i < data.length; i += 1) {
-            elevators.push(`${data[i].title}: ${data[i].description}`);
+    /**
+     * @function getAdditionalData
+     * @description Requests additional data for incidents lines/elevators.
+     * @async
+     *
+     * @param {string} url - URL to API endpoint
+     * @param {string} type - Incident type
+     *
+     * @returns {void}
+     */
+    async getAdditionalData(url, type) {
+        try {
+            const response = await fetch(url);
+            const parsedBody = await response.json();
+
+            if (Object.prototype.hasOwnProperty.call(parsedBody.data, 'trafficInfos')) {
+                this[`handle${type}Data`](parsedBody.data.trafficInfos);
+            } else {
+                throw new Error('No WienerLinien data');
+            }
+        } catch (e) {
+            console.log(`Error getting WienerLinien data for type ${type}:`, e);
         }
-        elevators.sort();
+    },
+
+    /**
+     * @function handleElevatorData
+     * @description Maps elevator response and sends it to the module via socket notification.
+     *
+     * @param {object[]} data - Response from API
+     *
+     * @returns {void}
+     */
+    handleElevatorData(data = []) {
+        const elevators = data
+            .map(({ title, description }) => `${title}: ${description}`)
+            .sort();
 
         this.sendSocketNotification('ELEVATORS', elevators);
     },
 
-    handleIncidentData(data) {
-        const incidents = [];
-        for (let i = 0; i < data.length; i += 1) {
-            const lines = data[i].relatedLines.join(', ');
-            const description = data[i].description;
-            incidents.push({ lines, description });
-        }
-        incidents.sort((a, b) => {
-            const nameA = a.lines.toUpperCase();
-            const nameB = b.lines.toUpperCase();
-            if (nameA < nameB) {
-                return -1;
-            } else if (nameA > nameB) {
-                return 1;
-            }
-            return 0;
-        });
+    /**
+     * @function handleIncidentData
+     * @description Maps incident response and sends it to the module via socket notification.
+     *
+     * @param {object[]} data - Response from API
+     *
+     * @returns {void}
+     */
+    handleIncidentData(data = []) {
+        const incidents = data
+            .map(({ relatedLines, description }) => ({
+                description,
+                lines: relatedLines.join(', ')
+            }))
+            .sort((a, b) => a.lines.localeCompare(b.lines));
 
         this.sendSocketNotification('INCIDENTS', incidents);
     },
 
+    /**
+     * @function handleData
+     * @description Maps station departure response and sends it to the module via socket notification.
+     *
+     * @param {object[]} data - Response from API
+     * @param {string} time - server time
+     *
+     * @returns {void}
+     */
     handleData(data, time) {
         const stations = {};
 
@@ -130,7 +185,7 @@ module.exports = NodeHelper.create({
                         break;
                     }
 
-                    const departureTime = data[i].lines[n].departures.departure[x].departureTime;
+                    const { departureTime } = data[i].lines[n].departures.departure[x];
 
                     stations[data[i].locationStop.properties.name].departures.push({
                         time: departureTime[Object.prototype.hasOwnProperty.call(departureTime, 'timeReal') ? 'timeReal' : 'timePlanned'],
@@ -141,10 +196,13 @@ module.exports = NodeHelper.create({
                 }
                 if (metroFlag) {
                     const departureTimePattern = /[0-9]+/g;
-                    const departureTimeMatches = data[i].lines[n].towards.match(departureTimePattern).toString().split(',');
+                    const departureTimeMatches = data[i].lines[n].towards.match(departureTimePattern).toString()
+                        .split(',');
 
                     const towardsPattern = /^[a-zäöüß ]+/i;
-                    const towardsMatch = data[i].lines[n].towards.match(towardsPattern).toString().replace(/ {2,}/g, ' ').trim();
+                    const towardsMatch = data[i].lines[n].towards.match(towardsPattern).toString()
+                        .replace(/ {2,}/g, ' ')
+                        .trim();
 
                     for (let x = 0; x < departureTimeMatches.length; x += 1) {
                         const datetime = new Date(time);
@@ -165,14 +223,7 @@ module.exports = NodeHelper.create({
         const keys = Object.keys(stations);
 
         for (let i = 0; i < keys.length; i += 1) {
-            stations[keys[i]].departures.sort((a, b) => {
-                if (a.time < b.time) {
-                    return -1;
-                } else if (a.time > b.time) {
-                    return 1;
-                }
-                return 0;
-            });
+            stations[keys[i]].departures.sort((a, b) => a.time - b.time);
         }
 
         this.sendSocketNotification('STATIONS', stations);